Massive refactor with minor forward-facing difference
Split all the stepMachine constructions into their own file.
New settings on the developer settings page.
The graph test is accessible from the developer page and now tests
  magnitude over long watch periods (for understanding how to
  implement sleep mode).
Step intensity (as seen in the squeaky shoes) is now determined from a
  maximum threshold (rather than just increasing by 4 for each sound).
The new default maximum threshold is 10 (this is the only outwardly-
  visible change: bike horn sounds are now slightly less difficult
  to trigger.)