Open external browser for all http(s):// links rather than target=_blank
Also don't return false from the event handler.

This is the behavior I actually want anyway, and as logn as I'm triggering it withth an event interception, I might as well code it like that.
Of course, it's a moot point right now, but it'll be nice if/when I actually fix the bug that's stopping this from working.